Implementirana logika za dobavljanje svih tura ulogovanog korisnika
