+++ app/tours/pages/guide-tours/guide-tours.html
@@ -1,26 +1,27 @@
 <!DOCTYPE html>
 <html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>My Tours</title>
-    <link rel="stylesheet" href="../../../../dist/tours/pages/guide-tours/guide-tours.css">
-</head>
-<body>
+    <head>
+        <meta charset="UTF-8">
+        <meta name="viewport" content="width=device-width, initial-scale=1.0">
+        <title>My Tours</title>
+        <link rel="stylesheet" href="../../../../dist/tours/pages/guide-tours/guide-tours.css">
+        <script type="module" src="/dist/tours/pages/guide-tours/guide-tours.js"></script>
+    </head>
+    <body>
 
-    <nav class="navbar">
-        <ul class="nav-list">
-            <li class="nav-item"><a href="../../../index.html">Home</a></li>
-            <li class="nav-item" id="login"><a href="login.html">Login</a></li>
-            <li class="nav-item" id="logout"><a href="../../../index.html">Logout</a></li>
-        </ul>
-    </nav>
-    <div class="container">
-        <h1>My Tours</h1>
+        <nav class="navbar">
+            <ul class="nav-list">
+                <li class="nav-item"><a href="../../../index.html">Home</a></li>
+                <li class="nav-item" id="login"><a href="login.html">Login</a></li>
+                <li class="nav-item" id="logout"><a href="../../../index.html">Logout</a></li>
+            </ul>
+        </nav>
+        <div class="container">
+            <h1>My Tours</h1>
 
-        <div id="toursList" class="tour-list"></div>
-    </div>
-
-    <script type="module" src="/dist/tours/pages/guide-tours/guide-tours.js"></script>
-</body>
+            <div id="toursList" class="tour-list"></div>
+        </div>
+        
+    </body>
 </html>
+
